MIAZ_practice/Practice_1.5/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.9
Flask-SQLAlchemy==2.5.1